Half-hearted support for structure.sql
Apartment doesn't support creation of tenants on Postgres using
tenants with sql schema format (struture.sql file). This is because
pg_dump creates a structure.sql file that includes CREATE SCHEMA and
other statements that are schema-specific.

This change assumes that structure.sql has been modified manually to
remove these statements so that structure.sql only describes and empty
structure of a new tenant.

It's an unelegant hack.